requirements.txt
flask
python-vlc